Add dualgun class a new type of weapon

+++ src/entity.cpp
@@ -32,8 +32,8 @@ void Entity::render(){
     SDL_Rect dst;
     dst.w = size.x;
     dst.h = size.y;
-    dst.x = pos.x;
-    dst.y = pos.y;
+    dst.x = (int)pos.x;
+    dst.y = (int)pos.y;
     SDL_RendererFlip flip = SDL_FLIP_NONE;
     if(isFlipping) flip = SDL_FLIP_HORIZONTAL;
     SDL_RenderCopyEx(gb::getRenderer(), tex, NULL, &dst, 0, NULL, flip);
